refactor: rename file with algolia to offer as it is only about offer
